+++ .gitignore
@@ -21,4 +21,9 @@ eggs/
 
 *.Vertex.json
 *.Edge.json
-cache.sqlite
+*.sqlite
+source/reference/*.json
+.coverage
+
+.cache/
+.pytest_cache/
