+++ src/bmeg/maf/gene_enricher.py
@@ -9,7 +9,7 @@ GENES = {}
 ALIASES = {}
 
 # get file if not already present
-fname = 'non_alt_loci_set.json'
+fname = 'source/reference/non_alt_loci_set.json'
 if not os.path.isfile(fname):
     url = 'ftp://ftp.ebi.ac.uk/pub/databases/genenames/new/json/non_alt_loci_set.json'  # noqa
     urllib.request.urlretrieve(url, fname)
