+++ transform/mc3/README.md
@@ -1,10 +1,3 @@
-
-
-ftp://ftp.ensembl.org/pub/grch37/release-90/fasta/homo_sapiens/dna/Homo_sapiens.GRCh37.dna.primary_assembly.fa.gz
-
-ftp://ftp.ensembl.org/pub/grch37/release-90/variation/VEP/homo_sapiens_vep_90_GRCh37.tar.gz
-
-
 ## verify source style
 ```
 flake8
