--- transform/variant/.gitignore
@@ -1,8 +1,0 @@
-.cache/
-.pytest_cache/
-.DS_Store
-.coverage
-biostream/
-source/
-*.log
-non_alt_loci_set.json
