--- transform/variant/tests/.gitignore
@@ -1,3 +1,0 @@
-test.*.Vertex.json
-test.*.Edge.json
-test.myvariant.info.sqlite
