--- transform/variant/tests/ccle_test.maf
@@ -1,10 +1,0 @@
-Hugo_Symbol	Entrez_Gene_Id	NCBI_Build	Chromosome	Start_position	End_position	Strand	Variant_Classification	Variant_Type	Reference_Allele	Tumor_Seq_Allele1	dbSNP_RS	dbSNP_Val_Status	Genome_Change	Annotation_Transcript	Tumor_Sample_Barcode	cDNA_Change	Codon_Change	Protein_Change	isDeleterious	isTCGAhotspot	TCGAhsCnt	isCOSMIChotspot	COSMIChsCnt	ExAC_AF	WES_AC	WGS_AC	SangerWES_AC	SangerRecalibWES_AC	RNAseq_AC	HC_AC	RD_AC
-AGRN	375790	37	1	979072	979072	+	Silent	SNP	A	G	NA	NA	g.chr1:979072A>G	ENST00000379370.2	22RV1_PROSTATE	c.1758A>G	c.(1756-1758)acA>acG	p.T586T	FALSE	FALSE	0	FALSE	0		27:24	15:13	9:10	9:12	104:20	NA	NA
-ATAD3A	55210	37	1	1459233	1459233	+	Silent	SNP	A	G	NA	NA	g.chr1:1459233A>G	ENST00000378755.5	22RV1_PROSTATE	c.1122A>G	c.(1120-1122)gcA>gcG	p.A374A	FALSE	FALSE	0	FALSE	0	8.237e-06	29:49	17:31	33:40	30:38		NA	NA
-NADK	65220	37	1	1685635	1685635	+	Missense_Mutation	SNP	G	A	NA	NA	g.chr1:1685635G>A	ENST00000341426.5	22RV1_PROSTATE	c.956C>T	c.(955-957)tCc>tTc	p.S319F	FALSE	FALSE	0	FALSE	0		25:39	14:23	16:19	17:20	176:266	NA	NA
-PLCH2	9651	37	1	2436128	2436128	+	Missense_Mutation	SNP	G	A	NA	NA	g.chr1:2436128G>A	ENST00000419816.2	22RV1_PROSTATE	c.3727G>A	c.(3727-3729)Gtg>Atg	p.V1243M	FALSE	FALSE	0	FALSE	0		9:20	23:15	19:22	20:20		NA	NA
-LRRC47	57470	37	1	3703695	3703695	+	Silent	SNP	G	A	NA	NA	g.chr1:3703695G>A	ENST00000378251.1	22RV1_PROSTATE	c.795C>T	c.(793-795)ggC>ggT	p.G265G	FALSE	FALSE	0	FALSE	0	3.295e-05	19:21	11:16	7:19	8:17	87:104	NA	NA
-CEP104	9731	37	1	3761503	3761503	+	Silent	SNP	G	A	NA	NA	g.chr1:3761503G>A	ENST00000378230.3	22RV1_PROSTATE	c.534C>T	c.(532-534)agC>agT	p.S178S	FALSE	TRUE	4	FALSE	0	8.236e-06	19:29	21:16	25:33	22:29	109:108	NA	NA
-CHD5	26038	37	1	6195391	6195391	+	Silent	SNP	C	T	NA	NA	g.chr1:6195391C>T	ENST00000262450.3	22RV1_PROSTATE	c.2769G>A	c.(2767-2769)ctG>ctA	p.L923L	FALSE	FALSE	0	FALSE	0		22:30	22:21	22:39	22:39		150:152	NA
-RERE	473	37	1	8716189	8716189	+	Silent	SNP	A	G	NA	NA	g.chr1:8716189A>G	ENST00000337907.3	22RV1_PROSTATE	c.168T>C	c.(166-168)agT>agC	p.S56S	FALSE	FALSE	0	FALSE	0		154:188	24:15	133:151	125:143	54:28	NA	NA
-RERE	473	37	1	8716245	8716245	+	Missense_Mutation	SNP	G	A	NA	NA	g.chr1:8716245G>A	ENST00000337907.3	22RV1_PROSTATE	c.112C>T	c.(112-114)Cgc>Tgc	p.R38C	FALSE	FALSE	0	FALSE	0		136:138	21:20	98:116	95:114	48:21	NA	NA
