--- transform/variant/tests/tcga_test-NO_REF_ALT.maf
@@ -1,2 +1,0 @@
-Hugo_Symbol	Entrez_Gene_Id	Center	NCBI_Build	Chromosome	Start_Position	End_Position	Strand	Variant_Classification	Variant_Type	Reference_Allele	Tumor_Seq_Allele1	Tumor_Seq_Allele2	dbSNP_RS	dbSNP_Val_Status	Tumor_Sample_Barcode	Matched_Norm_Sample_Barcode	Match_Norm_Seq_Allele1	Match_Norm_Seq_Allele2	Tumor_Validation_Allele1	Tumor_Validation_Allele2	Match_Norm_Validation_Allele1	Match_Norm_Validation_Allele2	Verification_Status	Validation_Status	Mutation_Status	Sequencing_Phase	Sequence_Source	Validation_Method	Score	BAM_File	Sequencer	Tumor_Sample_UUID	Matched_Norm_Sample_UUID	HGVSc	HGVSp	HGVSp_Short	Transcript_ID	Exon_Number	t_depth	t_ref_count	t_alt_count	n_depth	n_ref_count	n_alt_count	all_effects	Allele	Gene	Feature	Feature_type	Consequence	cDNA_position	CDS_position	Protein_position	Amino_acids	Codons	Existing_variation	ALLELE_NUM	DISTANCE	STRAND	SYMBOL	SYMBOL_SOURCE	HGNC_ID	BIOTYPE	CANONICAL	CCDS	ENSP	SWISSPROT	TREMBL	UNIPARC	RefSeq	SIFT	PolyPhen	EXON	INTRON	DOMAINS	GMAF	AFR_MAF	AMR_MAF	ASN_MAF	EAS_MAF	EUR_MAF	SAS_MAF	AA_MAF	EA_MAF	CLIN_SIG	SOMATIC	PUBMED	MOTIF_NAME	MOTIF_POS	HIGH_INF_POS	MOTIF_SCORE_CHANGE	IMPACT	PICK	VARIANT_CLASS	TSL	HGVS_OFFSET	PHENO	MINIMISED	ExAC_AF	ExAC_AF_AFR	ExAC_AF_AMR	ExAC_AF_EAS	ExAC_AF_FIN	ExAC_AF_NFE	ExAC_AF_OTH	ExAC_AF_SAS	GENE_PHENO	FILTER	COSMIC	CENTERS	CONTEXT	DBVS	NCALLERS
-ZFP3	0	.	GRCh37	17	4998659	4998659	+	3'UTR	SNP				rs776510613	.	TCGA-2W-A8YY-01A-11D-A37N-09	TCGA-2W-A8YY-10A-01D-A37N-09	G	G	..	.	.	.	.	.	.	.	.	.	.	.	.	.	c.*2351G>A	.	.	ENST00000318833	2/2	35	19	16	30	30	0ZFP3,3_prime_UTR_variant,,ENST00000318833,;	A	ENSG00000180787	ENST00000318833	Transcript	3_prime_UTR_variant	4196	.	.	.	.	rs776510613	.	.	1	ZFP3	HGNC	12861	protein_coding	YES	CCDS11067.1	ENSP00000320347	ZFP3_HUMAN	.	UPI0000070244	.	.	.	2/2	.	.	.	.	.	.	.	..	.	.	.	.	.	.	.	.	.	MODIFIER	.	SNV	.	.	.	.	.	.	.	.	.	.	.	.	.PASS	NONE	MUTECT|RADIA|SOMATICSNIPER|MUSE|VARSCANS	GCGCCGTCTTA	.	5
