--- transform/variant/tests/tcga_test-no-center.maf
@@ -1,2 +1,0 @@
-Hugo_Symbol	Entrez_Gene_Id	Foo	NCBI_Build	Chromosome	Start_Position	End_Position	Strand	Variant_Classification	Variant_Type	Reference_Allele	Tumor_Seq_Allele1	Tumor_Seq_Allele2	dbSNP_RS	dbSNP_Val_Status	Tumor_Sample_Barcode	Matched_Norm_Sample_Barcode	Match_Norm_Seq_Allele1	Match_Norm_Seq_Allele2	Tumor_Validation_Allele1	Tumor_Validation_Allele2	Match_Norm_Validation_Allele1	Match_Norm_Validation_Allele2	Verification_Status	Validation_Status	Mutation_Status	Sequencing_Phase	Sequence_Source	Validation_Method	Score	BAM_File	Sequencer	Tumor_Sample_UUID	Matched_Norm_Sample_UUID	HGVSc	HGVSp	HGVSp_Short	Transcript_ID	Exon_Number	t_depth	t_ref_count	t_alt_count	n_depth	n_ref_count	n_alt_count	all_effects	Allele	Gene	Feature	Feature_type	Consequence	cDNA_position	CDS_position	Protein_position	Amino_acids	Codons	Existing_variation	ALLELE_NUM	DISTANCE	STRAND	SYMBOL	SYMBOL_SOURCE	HGNC_ID	BIOTYPE	CANONICAL	CCDS	ENSP	SWISSPROT	TREMBL	UNIPARC	RefSeq	SIFT	PolyPhen	EXON	INTRON	DOMAINS	GMAF	AFR_MAF	AMR_MAF	ASN_MAF	EAS_MAF	EUR_MAF	SAS_MAF	AA_MAF	EA_MAF	CLIN_SIG	SOMATIC	PUBMED	MOTIF_NAME	MOTIF_POS	HIGH_INF_POS	MOTIF_SCORE_CHANGE	IMPACT	PICK	VARIANT_CLASS	TSL	HGVS_OFFSET	PHENO	MINIMISED	ExAC_AF	ExAC_AF_AFR	ExAC_AF_AMR	ExAC_AF_EAS	ExAC_AF_FIN	ExAC_AF_NFE	ExAC_AF_OTH	ExAC_AF_SAS	GENE_PHENO	FILTER	COSMIC	CENTERS	CONTEXT	DBVS	NCALLERS
-TACC2	0	.	GRCh37	10	123810032	123810032	+	Missense_Mutation	SNP	C	C	T	rs540521848	.	TCGA-02-0003-01A-01D-1490-08	TCGA-02-0003-10A-01D-1490-08	C	C	.	.	.	.	.	.	.	.	.	.	.	.	.	.	.	c.113C>T	p.Thr38Met	p.T38M	ENST00000369005	3/23	43	4	38	74	74	0	TACC2,missense_variant,p.Thr38Met,ENST00000358010,;TACC2,missense_variant,p.Thr38Met,ENST00000453444,;TACC2,missense_variant,p.Thr38Met,ENST00000513429,;TACC2,missense_variant,p.Thr55Met,ENST00000491540,;TACC2,missense_variant,p.Thr38Met,ENST00000369005,;TACC2,missense_variant,p.Thr38Met,ENST00000334433,;TACC2,missense_variant,p.Thr38Met,ENST00000515603,;TACC2,missense_variant,p.Thr38Met,ENST00000515273,;TACC2,non_coding_transcript_exon_variant,,ENST00000498721,;TACC2,non_coding_transcript_exon_variant,,ENST00000492689,;	T	ENSG00000138162	ENST00000369005	Transcript	missense_variant	453	113	38	T/M	aCg/aTg	rs540521848,COSM465374,COSM2148937	.	.	1	TACC2	HGNC	11523	protein_coding	YES	CCDS7626.1	ENSP00000358001	TACC2_HUMAN	.	UPI0000246F6B	.	.	benign(0.335)	3/23	.	.	T:0.0002	T:0	T:0.0014	.	T:0	T:0	T:0	.	.	.	0,1,1	.	.	.	.	.	MODERATE	.	SNV	.	.	0,1,1	.	.	.	.	.	.	.	.	.	.	PASS	SITE|p.T38M|c.113C>T|3	MUTECT|RADIA|SOMATICSNIPER|MUSE|VARSCANS	GGACACGCCCG	by1000G	5
